fix(ci): first-party line limit — extract observability value types
