+++ src/awf/service/workspace_observability.py
@@ -7,9 +7,8 @@ import re
 from collections.abc import Iterator, Mapping, Sequence
 from contextlib import contextmanager
 from contextvars import ContextVar
-from dataclasses import dataclass
 from datetime import UTC, datetime
-from typing import Any, Literal, Protocol, TypedDict, cast
+from typing import Any, cast
 
 from sqlalchemy.ext.asyncio import AsyncSession
 
@@ -39,16 +38,33 @@ from awf.service.bounded_list import (
 )
 from awf.service.coordination import coordination_warnings_from_task_policy
 from awf.service.profile_metadata import network_posture_from_profile_snapshot
-from awf.service.provider_recovery import (
-    ProviderRecoveryStateView,
-    provider_recovery_state_for_workspace,
-)
+from awf.service.provider_recovery import provider_recovery_state_for_workspace
 from awf.service.usage_store import (
     USAGE_SOURCE,
     UsageSnapshot,
     read_latest_usage_snapshot,
     read_latest_usage_snapshots,
 )
+from awf.service.workspace_observability_types import (
+    AgentIdentity,
+    AgentIdentityPayload,
+    AgentIdentitySource,
+    LifecycleStagePayload,
+    LifecycleStageSummary,
+    LlmUsagePayload,
+    LlmUsageSummary,
+    WorkspaceIdentityUsagePayload,
+    WorkspaceObservabilityPayload,
+    WorkspaceRecoveryCurrentOperation,
+    WorkspaceRecoveryPayload,
+    WorkspaceRecoverySummary,
+    _LifecycleAccumulator,
+    _RecoveryOperationLike,
+)
+from awf.service.workspace_observability_types import (
+    LifecycleStageStatus as LifecycleStageStatus,
+)
+from awf.service.workspace_observability_types import LlmUsageStatus as LlmUsageStatus
 from awf.service.workspace_overview_pagination import (
     InvalidWorkspaceOverviewCursorError as InvalidWorkspaceOverviewCursorError,
 )
@@ -62,9 +78,6 @@ from awf.service.workspace_overview_pagination import (
     _WorkspaceOverviewCursor as _WorkspaceOverviewCursor,
 )
 
-AgentIdentitySource = Literal["task_policy", "default", "unavailable"]
-LifecycleStageStatus = Literal["pending", "active", "completed", "terminal_skipped"]
-LlmUsageStatus = Literal["available", "unavailable"]
 _log = get_logger(__name__)
 
 DEFAULT_STALE_REASON_LIMIT = 50
@@ -137,139 +150,6 @@ _GENERIC_RECOVERY_REASON_CODES = frozenset(
 _MAX_RECOVERY_PAYLOAD_KEYS = 32
 _MAX_RECOVERY_PAYLOAD_DEPTH = 4
 _MAX_RECOVERY_PAYLOAD_SEQUENCE_ITEMS = 20
-
-
-@dataclass(frozen=True)
-class AgentIdentity:
-    model: str | None
-    effort: str | None
-    model_source: AgentIdentitySource
-    effort_source: AgentIdentitySource
-
-
-@dataclass(frozen=True)
-class LifecycleStageSummary:
-    stage: str
-    started_at: datetime | None
-    ended_at: datetime | None
-    duration_seconds: int | None
-    status: LifecycleStageStatus
-
-
-@dataclass(frozen=True)
-class LlmUsageSummary:
-    input_tokens: int | None
-    output_tokens: int | None
-    total_tokens: int | None
-    cost_estimate: float | None
-    currency: str | None
-    status: LlmUsageStatus
-    source: str
-    reason: str | None
-    cached_input_tokens: int | None = None
-    reasoning_output_tokens: int | None = None
-
-
-@dataclass(frozen=True)
-class WorkspaceRecoveryCurrentOperation:
-    id: str
-    type: str
-    status: str
-    created_at: datetime
-    started_at: datetime | None
-    payload: dict[str, Any] | None
-
-
-@dataclass(frozen=True)
-class WorkspaceRecoverySummary:
-    from_state: str | None
-    to_state: str | None
-    reason_code: str | None
-    action: str | None
-    recovery_mode: str | None
-    started_at: datetime
-    current_operation: WorkspaceRecoveryCurrentOperation | None
-    summary: str
-    payload: dict[str, Any] | None
-    started_event_order: int | None = None
-    provider_recovery: ProviderRecoveryStateView | None = None
-
-
-class AgentIdentityPayload(TypedDict):
-    agent_model: str | None
-    agent_effort: str | None
-    cursor_auto_mode: str | None
-    agent_model_source: AgentIdentitySource
-    agent_effort_source: AgentIdentitySource
-
-
-class LifecycleStagePayload(TypedDict):
-    stage: str
-    started_at: datetime | None
-    ended_at: datetime | None
-    duration_seconds: int | None
-    status: LifecycleStageStatus
-
-
-class LlmUsagePayload(TypedDict):
-    input_tokens: int | None
-    cached_input_tokens: int | None
-    output_tokens: int | None
-    reasoning_output_tokens: int | None
-    total_tokens: int | None
-    cost_estimate: float | None
-    currency: str | None
-    status: LlmUsageStatus
-    source: str
-    reason: str | None
-
-
-class WorkspaceRecoveryCurrentOperationPayload(TypedDict):
-    id: str
-    type: str
-    status: str
-    created_at: datetime
-    started_at: datetime | None
-    payload: dict[str, Any] | None
-
-
-class WorkspaceRecoveryPayload(TypedDict):
-    from_state: str | None
-    to_state: str | None
-    reason_code: str | None
-    action: str | None
-    recovery_mode: str | None
-    started_at: datetime
-    started_event_order: int | None
-    current_operation: WorkspaceRecoveryCurrentOperationPayload | None
-    summary: str
-    payload: dict[str, Any] | None
-    provider_recovery: dict[str, Any] | None
-
-
-class WorkspaceObservabilityPayload(AgentIdentityPayload):
-    lifecycle: list[LifecycleStagePayload]
-    llm_usage: LlmUsagePayload
-    recovery: WorkspaceRecoveryPayload | None
-
-
-class WorkspaceIdentityUsagePayload(AgentIdentityPayload):
-    llm_usage: LlmUsagePayload
-
-
-class _RecoveryOperationLike(Protocol):
-    id: object
-    type: object
-    status: object
-    payload: object
-    created_at: datetime
-    started_at: datetime | None
-
-
-@dataclass
-class _LifecycleAccumulator:
-    started_at: datetime | None = None
-    ended_at: datetime | None = None
 
 
 async def list_workspace_overview_response(
